Finally get radios to talk to each other
This includes several bugfixes and improvements that aren't big enough
to elicit explicit callouts here. The one exception is the reset logic
for the RFM69HCW breakout, which has to be done with an explicit output.
This is due to the design of the breakout, which has a one-way level
shifter in the wrong direction. Removing the pull-up only caused the
input on the level shifter to be high-Z, which somehow resulted in a
logic high on the output side, so the breakout was always held in reset.